+++ Tests/GUUnitsGeneratorConversionsTests/units/SignsTests.swift
@@ -7,16 +7,16 @@ final class SignsTests: XCTestCase {
     /// Test t
     func testT() {
         XCTAssertEqual(Signs.t.rawValue, "t")
-        XCTAssertEqual(Signs.t.numericType, .int)
-        XCTAssertEqual(Signs.t.type, NumericTypes.int.rawValue)
+        XCTAssertEqual(Signs.t.numericType, .long)
+        XCTAssertEqual(Signs.t.type, NumericTypes.long.rawValue)
         XCTAssertFalse(Signs.t.isFloatingPoint)
     }
 
     /// Test U
     func testU() {
         XCTAssertEqual(Signs.u.rawValue, "u")
-        XCTAssertEqual(Signs.u.numericType, .uint)
-        XCTAssertEqual(Signs.u.type, NumericTypes.uint.rawValue)
+        XCTAssertEqual(Signs.u.numericType, .ulong)
+        XCTAssertEqual(Signs.u.type, NumericTypes.ulong.rawValue)
         XCTAssertFalse(Signs.u.isFloatingPoint)
     }
